[feature] - participant details WIP
