+++ constant/Role.ts
@@ -1,33 +1,33 @@
-class Role {
-    constructor(name: String, description: String, emoji: String) {
+export class Role {
+    constructor(name: string, description: string, emoji: string) {
         this.name = name;
         this.description = description;
         this.emoji = emoji;
     }
-    name: String;
-    description: String;
-    emoji: String;
+    name: string;
+    description: string;
+    emoji: string;
 }
 
-export const Roles = [
-    new Role(
+export const Roles = {
+    bigSepnder: new Role(
         "Big Spender",
         "Paid upfront in more than 50% of activities",
         "💸"
     ),
-    new Role(
+    earlyPayer: new Role(
         "Early Payer",
         "Net balance is positive",
         "⏱️"
     ),
-    new Role(
+    evenSplitter: new Role(
         "Even Splitter",
         "Net balance is close to zero",
         "🤝"
     ),
-    new Role(
+    chillPayer: new Role(
         "Chill Payer",
         "You take your time, but it all works out in the end",
         "🧊"
     )
-]
+}
